@@ -8,28 +8,16 @@ const routes = [
     component: HomeView
   },
   {
-    path: '/about',
-    name: 'about',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/AboutView.vue')
-  },
-  {
     path: '/todos',
     name: 'todos',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/TodosView.vue')
+
+    component: () => import('../views/TodosView.vue')
   },
   {
-    path:'/todo/:id',
-    name:'todo single',
-        // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/TodosDetail.vue')
+    path: '/todo/:id',
+    name: 'todo single',
+
+    component: () => import('../views/TodosDetail.vue')
   }
 ]
 
